refactor to avoid calling a function from the rule unused_function
