+++ crates/jarl-core/src/package.rs
@@ -14,10 +14,11 @@ use crate::lints::base::duplicated_function_definition::duplicated_function_defi
     compute_duplicates_from_shared, scan_top_level_assignments,
 };
 use crate::lints::base::unused_function::unused_function::{
-    collect_files, compute_unused_from_shared, has_cpp_extension, scan_symbols,
+    collect_files, compute_unused_from_shared, has_cpp_extension,
 };
 use crate::namespace::{parse_namespace_exports, parse_namespace_imports};
 use crate::rule_set::Rule;
+use crate::utils::scan_symbols;
 
 /// Scope of a file within an R package, determining how its definitions
 /// are checked for unused functions.
